ADDED: Successful context switch between two tasks

+++ src/stdlib.h
@@ -32,7 +32,7 @@
 #else
 
 typedef unsigned char uint8_t;
-typedef unsigned int uint16_t;
+typedef unsigned short uint16_t;
 typedef unsigned long uint32_t;
 typedef unsigned long long uint64_t;
